requirements.txt
aiofiles==23.1.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
certifi==2024.7.4
chardet==4.0.0
click==7.1.2
dnspython==2.6.1
email-validator==2.0.0
fastapi~=0.115.8
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
httptools>=0.3.0,<0.7.0
httpx==0.24.1
idna==3.7
itsdangerous==1.1.0
Jinja2==3.1.4
MarkupSafe==2.0.1
promise==2.3
pydantic~=2.10.6
python-dotenv==0.17.1
python-multipart==0.0.18
PyYAML>=5.4.1,<6.1.0
requests==2.32.0
Rx==1.6.1
starlette==0.40.0
urllib3==1.26.19
uvicorn==0.13.4
watchgod==0.7
websockets==10.0
bcrypt~=4.2.1
PyJWT==2.10.1
SQLAlchemy~=2.0.38
typing_extensions~=4.12.2
passlib~=1.7.4
pytest~=8.3.4
psycopg2-binary~=2.9.10
databases~=0.9.0
SQLAlchemy-Utils~=0.41.2